refactor: share attachment-summary builder and tidy composer attachments
